+++ CryptDisk/SOURCE/ManagerNew/AppMemory.h
@@ -17,6 +17,6 @@ typedef Singleton<SecureHeap, InitSecHeap> AppMemory;
 
 inline boost::shared_array<char> AllocPasswordBuffer(size_t length)
 {
-	return boost::shared_array<char>(static_cast<char*>(AppMemory::instance().Alloc(length)),
+	return boost::shared_array<char>(static_cast<char*>(AppMemory::instance().Alloc(static_cast<DWORD>(length))),
 		boost::bind(&SecureHeap::Free, boost::ref(AppMemory::instance()), _1));
 }
